move getQuota() method to Connection class
