+++ tests/MailboxTest.php
@@ -146,16 +146,6 @@ final class MailboxTest extends AbstractTest
         static::assertFalse(isset($status->uidnext), 'uidnext shouldn\'t be set');
     }
 
-    public function testQuota()
-    {
-        static::markTestIncomplete('imap_get_quotaroot isn\'t supported by the current c-client library');
-
-        $quota = $this->mailbox->getQuota();
-
-        static::assertArrayHasKey('usage', $quota);
-        static::assertArrayHasKey('limit', $quota);
-    }
-
     public function testBulkSetFlags()
     {
         // prepare second mailbox with 3 messages
